Added handling of variable sequence length data with PyTorch
Modified the Deep SHAP explainer code to be able to use input data that has variable sequence length. The main points are the introduction of a parameter that allows adding arguments for the model's feedforward method (such as an array of the original sequence lengths of each sequence id) and a boolean that indicates whether the input data has variable sequence length (so as to know how to remove padding and properly apply the feedforward method). Currently, these modifications are only for the PyTorch framework.